requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
googlemaps==4.10.0
pandas==2.1.1
openpyxl==3.1.2
beautifulsoup4==4.12.2
selenium==4.12.0
webdriver_manager==4.0.1
openai==0.28.1
